interview-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interview-coaching-agent"
version = "0.1.0"
description = "Standalone LiveKit interview coaching voice agent"
requires-python = ">=3.10"

dependencies = [
    "livekit-agents[deepgram,google,openai,silero,turn-detector]~=1.4",
    "livekit-plugins-sarvam==1.5.5",
    "livekit-plugins-noise-cancellation~=0.2.5",
    "python-dotenv>=1.0",
    "asyncpg>=0.30.0",
    "boto3>=1.35.0",
    "av==16.1.0",
    "langfuse>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = ["E501"]
